+++ CHANGELOG.md
@@ -1,18 +1,31 @@
+1.0.0
+- Improve icon modal layout on mobile
+- Add "no icons found" message to icons grid
+- fix pagination
+- add various footer links
+
 0.0.11
 - Single db query per search to favor speed over relevance
+
 0.0.10
 - Move supabase project
+
 0.0.9
 - Rebuild icons
+
 0.0.8
 - Use pino for all logging
 - Create lint rule to enforce that console log methods are not used
+
 0.0.7
 - Update security headers
+
 0.0.6
 - Enforce that version changes between prod commits
+
 0.0.2
 - Add automated version management system that updates Doppler environment variables from changelog
 - Change `icon` table to use `build_id` instead of `version`
+
 0.0.1
 - Setup environment variable validation + injection